+++ .babelrc
@@ -1,0 +1,9 @@
+{
+  "presets": [["@babel/preset-env", {"modules": false}],
+  "@babel/react"],
+  "env": {
+    "test": {
+      "plugins": ["transform-es2015-modules-commonjs", "dynamic-import-node", "transform-class-properties"]
+    }
+  }
+}
