+++ package.json
@@ -10,23 +10,54 @@
     "start": "webpack-dev-server  --inline --content-base . --history-api-fallback",
     "build": "webpack",
     "build-prod": "webpack -p",
-    "test": "react-scripts test --env=jsdom",
+    "test": "jest",
     "eject": "react-scripts eject"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
+  "jest": {
+    "modulePaths": [
+      "./src",
+      "./node_modules"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "jsx"
+    ],
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/tests/__mocks__/fileMock.js",
+      "\\.(css|less)$": "<rootDir>/tests/__mocks__/fileMock.js"
+    },
+    "testEnvironment": "node",
+    "testPathIgnorePatterns": [
+      "<rootDir>/node_modules/",
+      "<rootDir>/src/"
+    ],
+    "setupFiles": ["<rootDir>/tests/setup.js"]
+  },
   "devDependencies": {
+    "@babel/cli": "^7.0.0-beta.40",
+    "@babel/core": "^7.0.0-beta.40",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0-beta.40",
     "autoprefixer": "^8.6.2",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-eslint": "8.2.3",
+    "babel-jest": "22.4.3",
+    "babel-loader": "^8.0.0-beta.0",
+    "babel-plugin-dynamic-import-node": "^2.1.0",
+    "babel-plugin-lodash": "^3.3.2",
+    "babel-plugin-react-transform": "^3.0.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
     "clean-webpack-plugin": "^0.1.19",
     "css-loader": "^0.28.11",
     "cssnano": "^3.10.0",
+    "enzyme": "^3.6.0",
+    "enzyme-adapter-react-16": "^1.5.0",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
@@ -41,13 +72,14 @@
     "webpack-dev-server": "^3.1.4"
   },
   "dependencies": {
+    "@babel/preset-es2015": "^7.0.0-beta.53",
     "@material-ui/core": "^1.2.2",
     "@material-ui/icons": "^1.1.0",
     "@types/markerclustererplus": "2.1.33",
-    "babel-preset-es2015": "^6.24.1",
     "chartist": "^0.11.0",
     "dotenv": "^6.0.0",
     "history": "4.7.2",
+    "jest": "^22.4.2",
     "material-expansion-panel": "^1.0.0",
     "material-ui-fields": "^1.0.2",
     "material-ui-table-edit": "^3.0.3",
@@ -67,11 +99,12 @@
     "react-nouislider": "2.0.1",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "1.1.4",
     "react-swipeable-views": "^0.12.14",
     "react-table": "^6.8.6",
     "react-tagsinput": "3.19.0",
+    "react-test-renderer": "^16.0.0-alpha.12",
     "redux": "^4.0.0",
+    "redux-mock-store": "^1.5.3",
     "redux-thunk": "^2.3.0"
   }
 }
