+++ src/components/Header/HeaderLinks.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import classNames from "classnames";
 import PropTypes from "prop-types";
-import { Manager, Target, Popper } from "react-popper";
+///import { Manager, Target, Popper } from "react-popper";
 
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
