+++ src/routes/dashboard.jsx
@@ -2,11 +2,12 @@
 import Dashboard from "@material-ui/icons/Dashboard";
 import Person from "@material-ui/icons/Person";
 import ContentPaste from "@material-ui/icons/ContentPaste";
-import AddShopping from '@material-ui/icons/AddShoppingCart';
+import GridOn from '@material-ui/icons/GridOn';
 import PaymentIcon from '@material-ui/icons/Payment';
 import HomeIcon from '@material-ui/icons/Home';
 import AddIcon from '@material-ui/icons/Add';
-import LibraryBooks from '@material-ui/icons/LibraryBooks';
+import Settings from '@material-ui/icons/Settings';
+import SdStorage from '@material-ui/icons/SdStorage';
 import AccountBalance from '@material-ui/icons/AccountBalanceWallet';
 import CreditCard from '@material-ui/icons/CreditCard';
 
@@ -18,7 +19,7 @@ import VendorList from "../views/Vendor/index.js";
 import AddVendor from "../views/Vendor/add.js";
 import ViewVendor from "../views/Vendor/view.js";
 import PurchaseOrder from "../views/PurchaseOrder/index.js";
-import AddPurchaseRequisition from "../views/PurchaseRequisition/Add/index.js";
+import AddPurchaseRequisition from "../views/PurchaseRequisition/add.js";
 import ListPurchaseRequisition from "../views/PurchaseRequisition/index.js";
 import Log from "../views/Log/index.js";
 import Registration from "../views/RegistrationPage/index.js";
@@ -32,68 +33,41 @@ import UpgradeToPro from "views/UpgradeToPro/UpgradeToPro.jsx"; */
 export const dashboardRoutes = [
   {
     path: "/dashboard",
-    sidebarName: "Dashboard",
-    navbarName: "Dashboard",
-    icon: HomeIcon,
     component: DashboardPage
   },
    {
     path: "/user",
-    sidebarName: "User Profile",
-    navbarName: "Profile",
-    icon: Person,
     component: UserProfile
   },
   {
     path: "/vendor/view/:id",
-    sidebarName: "Vendors List",
-    navbarName: "Vendors List",
-    icon: ContentPaste,
     component: ViewVendor
   },{
     path: "/vendor/type/:type",
-    sidebarName: "Vendors List",
-    navbarName: "Vendors List",
     component: VendorList
   },
   {
     path: "/vendor",
-    sidebarName: "Vendors List",
-    navbarName: "Vendors List",
     component: VendorList
   },
   {
     path: "/vendor/add",
-    sidebarName: "Vendor Details",
-    navbarName: "Vendor Details",
-    icon: AddIcon,
     component: AddVendor
   },
   {
     path: "/purchaserequisition/add",
-    sidebarName: "New Purchase Requisition",
-    navbarName: "New Purchase Requisition",
-    icon: AddIcon,
     component: AddPurchaseRequisition
   },
   {
     path: "/purchaserequisition",
-    name: "Purchase Requisition",
-    icon: CreditCard,
     component: ListPurchaseRequisition
   },
   {
     path: "/PurchaseOrder",
-    sidebarName: "Purchase Order",
-    navbarName: "Purchase Order",
-    icon:  AccountBalance,
     component: PurchaseOrder
   },
   {
     path: "/Log",
-    sidebarName: "Report",
-    navbarName: "Report",
-    icon:  LibraryBooks,
     component: Log
   },
   { redirect: true, path: "/", to: "/dashboard", navbarName: "Redirect" }
@@ -106,40 +80,119 @@ export const AdminMenu = [
     icon: DashboardIcon,
     component: Dashboard
   },
+ {
+   collapse: true,
+   path: "/budgets",
+   name: "Budgets",
+   state: "openBudgets",
+   icon: PaymentIcon,
+   views: [
+    {
+      path: "#",
+      name: "Annual Budget",
+      mini: "AB",
+      component: ListPurchaseRequisition
+    },
+    {
+      path: "#",
+      name: "Monthly Budget",
+      mini: "MB",
+      component: ListPurchaseRequisition
+    },{
+      path: "#",
+      name: "Supplementary Budget",
+      mini: "SB",
+      component: ListPurchaseRequisition
+    }
+   ]
+ },
+ {
+   collapse: true,
+   path: "/inventorymgt",
+   name: "Inventory Mgt.",
+   state: "openInventoryMgt",
+   icon: SdStorage,
+   views: [
+    {
+      path: "#",
+      name: "Store Requisition",
+      mini: "SR",
+      component: ListPurchaseRequisition
+    },
+    {
+      path: "#",
+      name: "Gate Pass",
+      mini: "GP",
+      component: ListPurchaseRequisition
+    }
+   ]
+ },
+ {
+  collapse: true,
+  path: "/purchasing",
+  name: "Purchasing",
+  state: "openComponents",
+  icon: GridOn,
+  views: [
+    {
+      path: "/purchaserequisition",
+      name: "Purchase Requisition",
+      mini: "PR",
+      component: ListPurchaseRequisition
+    }, 
+    {
+      path: "/PurchaseOrder",
+      name: "Purchase Order",
+      mini: "PO",
+      component: PurchaseOrder
+    },    
+    {
+      path: "/RequestForQuotation",
+      name: "Purchase Order",
+      mini: "RFQ",
+      component: PurchaseOrder
+    }]
+  },
   {
-   path: "/user",
-   name: "User Profile",
-   icon: Person,
-   component: UserProfile
- },
- {
-   path: "/vendor",
-   name: "Vendors List",
-   icon: ContentPaste,
-   component: VendorList
- },
- {
-   path: "/vendor/add",
-   name: "Vendor Details",
-   icon: AddIcon,
-   component: AddVendor
- },
- {
-   path: "/purchaserequisition",
-   name: "Purchase Requisition",
-   icon: CreditCard,
-   component: ListPurchaseRequisition
- },
- {
-   path: "/PurchaseOrder",
-   name: "Purchase Order",
-   icon:  AccountBalance,
-   component: PurchaseOrder
- },
+    collapse: true,
+    path: "/qualitymgt",
+    name: "Quality Mgt.",
+    state: "openQualityMgt",
+    icon: ContentPaste,
+    views: [
+          {
+            path: "/vendor",
+            name: "Vendors Mgt",
+            mini: "VM",
+            component: VendorList
+          },
+    ]
+  }, 
+   {
+    collapse: true,
+    path: "/salesmgt",
+    name: "Sales Management",
+    state: "openSalesMgt",
+    icon: ContentPaste,
+    views: [
+     {
+       path: "#",
+       name: "Sales Order",
+       mini: "SO",
+       component: ListPurchaseRequisition
+     },
+     {
+       path: "#",
+       name: "Quotation",
+       mini: "Q",
+       component: ListPurchaseRequisition
+     }
+    ]
+  },
  {
    path: "/Log",
-   name: "Report",
-   icon:  LibraryBooks,
+   name: "Setup",
+   icon:  Settings,
    component: Log
  },
 ];
@@ -153,15 +206,15 @@ export const IacMenu = [
   {
     path: "/vendor/type/approved",
     name: "Approved Vendors",
-    icon: LibraryBooks,
+    icon: GridOn,
   },{
     path: "/vendor/type/pending",
     name: "Pending Approval",
-    icon: LibraryBooks,
+    icon: GridOn,
   },{
     path: "/vendor/type/blacklisted",
     name: "Blacklisted Vendors",
-    icon: LibraryBooks,
+    icon: GridOn,
   },
 ];
