--- src/test.js
@@ -1,1 +1,0 @@
-export default 5;
