+++ tests/setup.js
@@ -1,0 +1,4 @@
+var enzyme = require('enzyme');
+var Adapter = require('enzyme-adapter-react-16');
+
+enzyme.configure({ adapter: new Adapter() });
